updated AST PARSER WORKS FOR EXPR

+++ src/main/java/aar226_akc55_ayc62_ahl88/ast/Num.java
@@ -3,11 +3,11 @@ package aar226_akc55_ayc62_ahl88.ast;
 import aar226_akc55_ayc62_ahl88.ast.Expr;
 
 public class Num implements Expr {
-    int value;
-    public Num(int n) {
+    long value;
+    public Num(long n) {
         value = n;
     }
     public String toString() {
-        return Integer.toString(value);
+        return Long.toString(value);
     }
 }
